finished parser, stable version

+++ Sphere.h
@@ -3,11 +3,11 @@
 
 struct Point 
 {
-	float x;
-	float y;
-	float z;
+	double x;
+	double y;
+	double z;
 
-	Point(float x, float y, float z) 
+	Point(double x, double y, double z) 
 	{
 		this->x = x;
 		this->y = y;
@@ -18,13 +18,13 @@ struct Point
 class Sphere 
 {
 	Point center;
-	float radius;
+	double radius;
 	int lat;
 	int lon;
 	std::vector<Point> line_start;
 	std::vector<Point> line_end;
 public:
-	Sphere(Point c, float rad) : center(c), radius(rad), lat(16), lon(8) { calculateWireframe(); }
+	Sphere(Point c, double rad) : center(c), radius(rad), lat(16), lon(8) { calculateWireframe(); }
 	Sphere(const Sphere& s);
 	void calculateWireframe();
 	std::vector<Point>& getStartPoints() { return line_start; }
